# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quarkmed Workbench API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Add Model",
        "operationId": "add_model_api_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Edit Model",
        "operationId": "edit_model_api_models__model_id__put",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/test": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Test Model",
        "operationId": "test_model_api_models__model_id__test_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "gemini",
              "ollama"
            ],
            "title": "Provider"
          },
          "base_url": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Base Url"
          },
          "model_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "clear_key": {
            "type": "boolean",
            "title": "Clear Key",
            "default": false
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "supports_images": {
            "type": "boolean",
            "title": "Supports Images",
            "default": false
          },
          "default_text": {
            "type": "boolean",
            "title": "Default Text",
            "default": false
          },
          "default_image": {
            "type": "boolean",
            "title": "Default Image",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider",
          "base_url",
          "model_name"
        ],
        "title": "ModelIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}